Add correct starting dates and correct day

+++ Models/TJ.cs
@@ -1,7 +1,7 @@
 namespace TJ_API.Models;
 
 public record TJ(
-    int Days,
+    double Days,
     double Weeks,
     double Months,
     double Seconds
